refactor(p2): extract CouncilTaskStepView into sibling (stage 3 of 4)
Moves the Model Council UI system from ChatMessageTools.kt L1952-L2605
(parent ae597a0d, = end of file) into a new sibling
ChatMessageCouncilStep.kt. Bodies are byte-equal with the parent commit.

In the stage-2 coordinator (0cf4ae46) the same block had shifted to L1192-L1845;
stripped L1191-L1845 (leading blank + block) to keep coordinator EOF clean.

Stage ordering note (per pre-cut §6.1): Council extracts BEFORE SubAgent
because the SubAgent block contains calls into helpers physically declared
inside the Council block — `payloadObject` (parent L2521, called from
SubAgent block L1656), `extractFinalSubAgentText` (parent L2547, called from
SubAgent block L1817), and `extractSubAgentDisplayTextFromTranscript` (parent
L2569 area, called from SubAgent block L1816). Council goes first so these
helpers live in the Council sibling as `internal`, and the SubAgent block
(still in the coordinator at this stage) references them cross-file via the
same package.

Widening in sibling (private → internal):
  - payloadObject (used by SubAgent block, still in coord)
  - extractFinalSubAgentText (used by SubAgent block, still in coord)
  - extractSubAgentDisplayTextFromTranscript (used by SubAgent block, still
    in coord) — missed in pre-cut §3.2/§4 widening table; caught at compile
    gate.

Widening in coord (private → internal):
  - AgentToolStatus (enum class L142) — Council sibling uses it 5 times.
    Pre-cut §4 noted this gap (top-level private object/enum decls missing
    from pre-cut §1 table).
  - AgentToolKind (enum class L150) — Council sibling uses it 1 time.
  - AgentToolCallCapsule (fun L372) — Council sibling at L1995 of parent.
  - PHASE_LABEL_INTERVAL_MS (const val L964) — Council sibling at L1971 of
    parent. The const physically lives in the SubAgent block but the Council
    sibling references it cross-section; stays internal when relocated to
    SubAgent sibling in stage 4.

Coordinator strip: 1845 → 1190 (-655 body+blank) → 1180 (-10 import cleanup).
Imports trimmed: 10 this-refactor orphans (heightIn, ScrollableTabRow, Tab,
FontWeight, Dispatchers, withContext, JsonArray, ModelCouncilManager,
ModelCouncilRolePresets, readSubAgentDisplayTextFromTranscript). The
FontWeight + JsonArray refs partially spanned stage 1/2 ranges but the import
stayed live until Council's last reference left coord at this stage.

Pre-existing dead imports preserved (CLAUDE.md §3): Box, BoxWithConstraints,
CircleShape, FilledTonalButton, FilterChip, LocalContentColor, Color, getValue,
setValue (last two used implicitly via `by` — 8 sites in stripped coord).

Six `by` delegations in extracted Council body — sibling imports
androidx.compose.runtime.getValue + setValue explicitly per playbook §5.3.

Baseline 467 tests / 3 failed preserved.
